backend/requirements.txt
# --- Web Framework ---
Flask==3.0.0

# --- Data Handling ---
pandas==2.2.2
numpy==1.26.4

# --- Machine Learning ---
scikit-learn==1.5.0
joblib==1.4.2

# --- Code Metrics Analysis ---
radon==6.0.1

# --- Visualization (optional but useful for research analysis) ---
matplotlib==3.9.0
seaborn==0.13.2

# --- Development Utilities ---
Werkzeug==3.0.1

radon

pylint


google-generativeai